Extract Standards

Separates specified serial dilutions for fitting, background, and samples

extractStd(MFI, stdstr, bgstr, stddil, smpdil, antigen, fname, pdate)

Arguments

MFI

data frame with variable "Sample" that contains information about standard concentrations in "1/dilution" format (e.g. "1/200").

stdstr

character string indicating standards in the file's "Sample" column. Not case sensitive. If "" (empty string), standards will be determined by the pattern "1/" only.

bgstr

character string indicating background in the file's "Sample" column. Not case sensitive.

stddil

standard dilutions to use for standard curve fitting. If NULL, all the dilutions indicated in a "Sample" variable are used.

smpdil

single value for sample dilutions (if dilutions are not provided in smp data frame). Ignored if dilvar is provided and the variable is included in smp data frame.

antigen

character string.

fname

name of the file that contains raw data.

pdate

date of the plate processing.

Value

A list with standards for fitting, background values, sample values, indices for samples.

Details

details