Reformat raw luminex files for processing

Read in raw luminex files, extract standards and background (if provided), and format samples for processing.

prepLum(
  antigen,
  fname,
  fdir = NULL,
  dtype = "Median",
  stdstr = "std|stand",
  bgstr = "blank|background",
  stddil = NULL,
  smpdil = 1000,
  nwells = NULL,
  nsep = 2,
  ncolmax = 105,
  dformat = "%m/%d/%Y"
)

Arguments

antigen

character string.

fname

name of the file that contains raw data.

fdir

directory where the file is located (alternatively, full path can be included in fname).

dtype

character string for data type in the file.

stdstr

character string indicating standards in the file's "Sample" column. Not case sensitive. If "" (empty string), standards will be determined by the pattern "1/" only.

bgstr

character string indicating background in the file's "Sample" column. Not case sensitive.

stddil

a vector of standard dilutions. If NULL, dilutions are inferred from the file's "Sample" column. Non-null value can be used to exclude some dilutions from model fitting.

smpdil

single value for sample dilutions (if dilutions are not provided in smp data frame). Ignored if dilvar is provided and the variable is included in smp data frame.

nwells

number of wells. If NULL, inferred from the file.

nsep

number of lines separating different data types in the file.

ncolmax

maximum number of columns in the file.

dformat

date format in the file.

Value

A list with standards for fitting, background values, sample values, indices for samples.

Details

to be added