Plot the fit for standards and the samples

Produces a plot that includes points for standards, proposed fit, removed outliers, bounds for "flat" portions of the curve, and values for samples and for the background.

Usage

plotFit(
  std,
  xvar,
  yvar,
  fitpar = NULL,
  FUNmod = NULL,
  FUNinv = NULL,
  iout = NULL,
  bg = NULL,
  vsmp = NULL,
  smpflag = NULL,
  trimval = NULL,
  trimext = NULL,
  extrapolate.low = FALSE,
  extrapolate.up = FALSE,
  ylim = NULL,
  tcklab = NULL,
  stdcol = c("firebrick3", "darkslategray"),
  rugcol = c("cadetblue", "purple", "firebrick2"),
  ...
)

Arguments

std

matrix or data frame with standards for fitting.

xvar, yvar

character strings for the variables used to fit a standard curve. If NULL, first two columns are assumed to be x and y variables.

fitpar

values of function parameters.

FUNmod

model function.

FUNinv

model inverse function.

iout

indices of removed standard points.

bg

background values.

vsmp

sample values.

smpflag

character vector, flags for each sample.

trimval

for final results, the values at which the samples are trimmed.

trimext

integer vector of length two indicating if the values are trimmed at the extremum (lower and upper).

extrapolate.low

if TRUE, sample values beyond lower bounds will be processed by extrapolation of the standard curve (not recommended). Takes precedence over trim.flat value.

extrapolate.up

if TRUE, sample values beyond upper bounds will be processed by extrapolation of the standard curve (not recommended). Takes precedence over trim.flat value.

ylim

limits of the y-axis.

tcklab

tick labels for x-axis.

stdcol

vector of two colors for the standard points and the fit on the plot.

rugcol

vector of three colors for the rugplot, which indicates sample values (inside the bounds, between the bounds and extrema, and beyond extrema).

...

further graphical parameters.

Details

to be added